Modularized CNS
